Magnetic field visualization as surface
Use Axes3D.plot_surface() to visualize the points with equal magnitude
of the magnetic field, caused by the movement of a single charge. The
direction of the B vector at these points is not the same, but it rotates
along the direction of charge movement.

An experimental gradient B vectors are calculated, by using partial
derivatives of the B's magnitude, along X and Y/Z axes.