First version to calculate "real" EMI gradients
Calculate EMI gradient vectors from the partial derivatives of the
magnetic field, then integrate along the line segment that generates the
field. The derivatives are based on the field magnitude, the direction
(vector components) is ignored.

Fix the EMF direction, must be "v x B" instead of "B x v" (v is gradB).