Initial version of surface from normals util
Experimental tool to generate a surface based on a vector field of
normals to that surface. The algorithm consecutively approximates the
points, adjacent to the previously calculated border points. It starts
from a given base point.

The calculations are optimized to process the points from each of the 4
borders at once, by using numpy array operations.

TODO: Must perform iterations to increase the approximation precision.