Experimental point adjustment based on curvature
In surface from normals tool, improve the point adjustment, based on the
angle between normals. The distance between the surface points is
larger, where the curvature is smaller.

This is still not working as expected, esp. with non-trivial orientation
of the normals. The curvature is considered in a single dimension only.