Initial 3D draggable gizmo version
Experimental draggable module to allow dragging of an matlotlib 'quiver'
actor in '3d' projection axes. The actor is moved along the 'quiver'
direction only. The intention is to have 3 separate vectors for each
coordinate component.

TODO: The offset where the actor is dragged is still not coverted to its
coordinates.