Improve the self-inductance calculation hack
Locate and adjust all target points, that are collinear to any of the
source lines. The adjustment is a shift by some negligible distance,
currently this is 'sys.float_info.epsilon', but scaled to avoid float
precision losses. This repeats, till no such points.

Note: In self-induction mode, a huge Hall effect forces will be
generated, if 'tgt_coef != 0'. This is because of the above adjustment,
the points will be very close to a source line, with a very high B-s.